Store original event start date
The correction when the recurrence pattern leads to a non-existent date (i.e. Jan 30 => Feb 30, changes to Feb 28) means that subsequent events are shifted in date from the original event (i.e, Jan 30 => Feb 28 => Mar 28). The code is changed to add a _ctc_event_origin_date field to capture the original intent of the recurrence. This allows subsequent events to be corrected (i.e., Jan 30 => Feb 28 => Mar 30).